chore: storybook 에서 useNaviage 사용가능하게 해주는 부분 다른 방식으로 변경
