+++ package.json
@@ -24,7 +24,6 @@
     "react-icons": "^5.3.0",
     "react-router-dom": "^6.28.0",
     "react-select": "^5.8.3",
-    "storybook-addon-react-router-v6": "^2.0.15",
     "swiper": "^11.1.15",
     "vite-tsconfig-paths": "^5.1.3",
     "zustand": "^5.0.1"
